+++ src/main/java/ru/dromran/testtz/repository/PostEntityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import ru.dromran.testtz.entity.PostEntity;
 
-public interface PostEntityRepository extends JpaRepository<PostEntity, Integer>, JpaSpecificationExecutor<PostEntity> {
+public interface PostEntityRepository extends JpaRepository<PostEntity, Long>, JpaSpecificationExecutor<PostEntity> {
 
 }
